tidy api; create auth middleware
