+++ api/handle/comment.go
@@ -1,0 +1,58 @@
+package handle
+
+import (
+	"encoding/json"
+	"log"
+	"net/http"
+	"strconv"
+
+	"github.com/alacine/video_server/api/dbops"
+	"github.com/alacine/video_server/api/defs"
+	"github.com/alacine/video_server/api/utils"
+	"github.com/julienschmidt/httprouter"
+)
+
+func PostComment(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+	cbody := &defs.NewComment{}
+	dec := json.NewDecoder(r.Body)
+	if err := dec.Decode(cbody); err != nil {
+		log.Printf("(Error) PostComment: %s", err)
+		sendErrorResponse(w, defs.ErrorInternalFaults) // 500
+		return
+	}
+	vid, err := strconv.Atoi(p.ByName("vid"))
+	if err != nil {
+		log.Printf("(Error) PostComment: %s", err)
+		sendErrorResponse(w, defs.ErrorRequestBodyParseFailed) // 400
+	}
+	log.Printf("postcomment %v, %v", vid, cbody)
+	if err := dbops.AddNewComment(vid, cbody.AuthorId, cbody.Content); err != nil {
+		log.Printf("(Error) PostComment: %s", err)
+		sendErrorResponse(w, defs.ErrorDBError) // 500
+	} else {
+		sendNormalResponse(w, "ok", http.StatusCreated) // 201
+	}
+}
+
+func ListComments(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+	vid, err := strconv.Atoi(p.ByName("vid"))
+	if err != nil {
+		log.Printf("(Error) GetVideoInfo: %s", err)
+		sendErrorResponse(w, defs.ErrorRequestBodyParseFailed) /// 400
+		return
+	}
+	comments, err := dbops.ListComments(vid, 0, utils.GetCurrentTimestampSec())
+	if err != nil {
+		log.Printf("(Error) ListComments: %s", err)
+		sendErrorResponse(w, defs.ErrorDBError) // 500
+		return
+	}
+	cs := &defs.Comments{Comment: comments}
+	if resp, err := json.Marshal(cs); err != nil {
+		log.Printf("(Error) ListComments: %s", err)
+		sendErrorResponse(w, defs.ErrorInternalFaults) // 500
+		return
+	} else {
+		sendNormalResponse(w, string(resp), http.StatusOK) /// 200
+	}
+}
