+++ api/handle/response.go
@@ -1,4 +1,4 @@
-package main
+package handle
 
 import (
 	"encoding/json"
